+++ src/electron/main.ts
@@ -1,9 +1,10 @@
 import path from "path";
-import { animateWindowTransition, isDev } from "./util.js";
-import { app, BrowserWindow, ipcMain, screen } from "electron";
-import { resolvePreloadPath } from "./path-resolver.js";
-import { exposeSlackApiMethods } from "./services/slack/expose-slack-api-methods.js";
 import dotenv from "dotenv";
+import { app, BrowserWindow, screen } from "electron";
+import { resolvePreloadPath } from "./path-resolver.js";
+import { animateWindowTransition, isDev } from "./util.js";
+import { exposeStore } from "./services/db/expose-store.js";
+import { exposeSlackApiMethods } from "./services/slack/expose-slack-api-methods.js";
 
 dotenv.config();
 
@@ -11,13 +12,14 @@ app.whenReady().then(() => {
   // app.dock?.hide();
   createWidgetWindow();
   exposeSlackApiMethods();
+  exposeStore();
 });
 
 function createWidgetWindow() {
   const { width: screenWidth, height: screenHeight } =
     screen.getPrimaryDisplay().bounds;
 
-  const initialWidgetWidth = 400;
+  const initialWidgetWidth = 1000;
   const initialWidgetHeight = 30;
   const activeWidgetHeight = screenHeight / 2 + 300;
   const paddingRight = 10;
@@ -45,7 +47,7 @@ function createWidgetWindow() {
   });
   if (isDev()) {
     mainWindow.loadURL("http://localhost:5173");
-    // mainWindow.webContents.openDevTools();
+    mainWindow.webContents.openDevTools();
   } else {
     mainWindow.loadFile(path.join(app.getAppPath(), "/dist-ui/index.html"));
   }