+++ src/shared/constants/electron-api-events.ts
@@ -1,0 +1,10 @@
+export const GET_FROM_STORE = "get-from-store";
+export const SET_IN_STORE = "set-in-store";
+export const DELETE_FROM_STORE = "delete-from-store";
+export const GET_STORE = "get-store";
+
+export const SLACK_INIT_AUTH = "slack:init-auth";
+export const SLACK_GET_AUTH = "slack:get-auth";
+export const GET_USER = "slack:get-user";
+export const SLACK_GET_CONVERSATIONS = "slack:get-conversations";
+export const SLACK_GET_CONVERSATION_HISTORY = "slack:get-conversation-history";