+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Merify
 
+## 0.7.0
+
+### Minor Changes
+
+- Add store management methods
+- Encrypt store data
+
 ## 0.6.0
 
 ### Minor Changes